+++ src/manager_api.rs
@@ -1,5 +1,4 @@
 use std::ops::Range;
-use std::panic;
 
 use itertools::Itertools;
 use rand::Rng;
@@ -82,51 +81,6 @@ pub fn chunk_to_global(pos: ChunkPos) -> IVec2 {
     atom.y += pos.chunk.y * CHUNK_LENGHT as i32;
 
     atom
-}
-
-/// Transforms global 3x3 chunk position to local 3x3 chunks position
-/// Used for chunk multithreaded updates
-pub fn global_to_local(pos: IVec2) -> (IVec2, i32) {
-    let range = 0..CHUNK_LENGHT as i32 * 3;
-    if !range.contains(&pos.x) || !range.contains(&pos.y) {
-        panic!("Invalid position on global_to_local.")
-    }
-
-    let chunk_lenght = CHUNK_LENGHT as i32;
-
-    let chunk_x = pos.x % (chunk_lenght * 3) / chunk_lenght;
-    let chunk_y = pos.y / chunk_lenght;
-
-    let local_x = pos.x - chunk_x * chunk_lenght;
-    let local_y = pos.y - chunk_y * chunk_lenght;
-
-    let chunk_index = chunk_y * 3 + chunk_x;
-
-    (IVec2::new(local_x, local_y), chunk_index)
-}
-
-/// Transforms local 3x3 chunk position to global 3x3 chunks position
-/// Used for chunk multithreaded updates
-pub fn local_to_global(pos: (IVec2, i32)) -> IVec2 {
-    let range = 0..CHUNK_LENGHT as i32;
-    if !range.contains(&pos.0.x) || !range.contains(&pos.0.y) || !(0..9).contains(&pos.1) {
-        panic!(
-            "Invalid position on local_to_global. {} {} {}",
-            pos.0.x, pos.0.y, pos.1
-        )
-    }
-
-    let chunk_size = CHUNK_LENGHT as i32;
-
-    let chunk_index = pos.1;
-
-    let chunk_x = chunk_index % 3;
-    let chunk_y = chunk_index / 3;
-
-    let global_x = pos.0.x + chunk_size * chunk_x;
-    let global_y = pos.0.y + chunk_size * chunk_y;
-
-    IVec2::new(global_x, global_y)
 }
 
 /// See if position is swapable, that means it sees if the position is a void
@@ -285,116 +239,6 @@ impl D1 for UVec2 {
     }
 }
 
-pub fn split_left_right(
-    array: &mut [Atom],
-) -> ([&mut Atom; CHUNK_LEN / 2], [&mut Atom; CHUNK_LEN / 2]) {
-    let (left, right): (Vec<_>, Vec<_>) = array
-        .chunks_mut(CHUNK_LENGHT)
-        .flat_map(|chunk| {
-            let (left, right) = chunk.split_at_mut(HALF_CHUNK_LENGHT);
-            left.iter_mut().zip(right.iter_mut()).collect::<Vec<_>>()
-        })
-        .unzip();
-
-    (left.try_into().unwrap(), right.try_into().unwrap())
-}
-
-pub fn updown_to_leftright(
-    array: &mut [Atom],
-) -> ([&mut Atom; CHUNK_LEN / 4], [&mut Atom; CHUNK_LEN / 4]) {
-    let (left, right): (Vec<_>, Vec<_>) = array
-        .chunks_mut(CHUNK_LENGHT)
-        .flat_map(|chunk| {
-            let (left, right) = chunk.split_at_mut(HALF_CHUNK_LENGHT);
-            left.iter_mut().zip(right.iter_mut()).collect::<Vec<_>>()
-        })
-        .unzip();
-
-    (left.try_into().unwrap(), right.try_into().unwrap())
-}
-
-//This splits up our chunks for the update step, while also mutably borrowing them, making a `ChunkReference`
-//Some chunks are not chopped others are chopped up/down, left/right, and also in four corners.
-//We do this because each center chunk needs half of the adjacent chunks
-//So it needs up/down/left/right halves, and also four corners
-//TODO Decrease individual atoms iterations to get a &mut Atom;
-pub fn get_mutable_references<'a>(
-    chunks: &'a mut HashMap<IVec2, Chunk>,
-    mutable_references: &mut HashMap<IVec2, ChunkReference<'a>>,
-    (x_toff, y_toff): (i32, i32),
-    dirty_rects: &HashMap<IVec2, URect>,
-    manager_pos: IVec2,
-) {
-    chunks
-        .iter_mut()
-        .filter(|(chunk_pos, _)| {
-            let same_x = (chunk_pos.x + x_toff + manager_pos.x.abs() % 2) % 2 == 0;
-            let same_y = (chunk_pos.y + y_toff + manager_pos.y.abs() % 2) % 2 == 0;
-            let step_as_center = same_x && same_y;
-            if step_as_center && dirty_rects.contains_key(*chunk_pos) {
-                return true;
-            } else if !step_as_center {
-                let to_check = match (same_x, same_y) {
-                    (false, false) => vec![ivec2(-1, -1), ivec2(-1, 1), ivec2(1, -1), ivec2(1, 1)],
-                    (true, false) => vec![ivec2(0, -1), ivec2(0, 1)],
-                    (false, true) => vec![ivec2(-1, 0), ivec2(1, 0)],
-                    _ => unreachable!(),
-                };
-                for vec in to_check {
-                    if dirty_rects.contains_key(&(**chunk_pos + vec)) {
-                        return true;
-                    }
-                }
-            }
-
-            false
-        })
-        .for_each(|(chunk_pos, chunk)| {
-            let same_x = (chunk_pos.x + x_toff + manager_pos.x.abs() % 2) % 2 == 0;
-            let same_y = (chunk_pos.y + y_toff + manager_pos.y.abs() % 2) % 2 == 0;
-
-            match (same_x, same_y) {
-                (true, true) => {
-                    mutable_references.insert(*chunk_pos, ChunkReference::Center(&mut chunk.atoms));
-                }
-                (true, false) => {
-                    let (up, down) = chunk.atoms.split_at_mut(CHUNK_LEN / 2);
-
-                    mutable_references.insert(
-                        *chunk_pos,
-                        ChunkReference::Side([
-                            Some(up.iter_mut().collect::<Vec<_>>().try_into().unwrap()),
-                            Some(down.iter_mut().collect::<Vec<_>>().try_into().unwrap()),
-                        ]),
-                    );
-                }
-                (false, true) => {
-                    let (left, right) = split_left_right(&mut chunk.atoms);
-
-                    mutable_references
-                        .insert(*chunk_pos, ChunkReference::Side([Some(left), Some(right)]));
-                }
-
-                (false, false) => {
-                    let (up, down) = chunk.atoms.split_at_mut(CHUNK_LEN / 2);
-
-                    let (up_left, up_right) = updown_to_leftright(up);
-                    let (down_left, down_right) = updown_to_leftright(down);
-
-                    mutable_references.insert(
-                        *chunk_pos,
-                        ChunkReference::Corner([
-                            Some(up_left),
-                            Some(up_right),
-                            Some(down_left),
-                            Some(down_right),
-                        ]),
-                    );
-                }
-            }
-        });
-}
-
 pub fn update_dirty_rects(dirty_rects: &mut HashMap<IVec2, URect>, pos: ChunkPos) {
     if let Some(dirty_rects) = dirty_rects.get_mut(&pos.chunk) {
         extend_rect_if_needed(dirty_rects, &pos.atom)
@@ -521,15 +365,6 @@ impl ChunkPos {
     pub fn new(atom: UVec2, chunk: IVec2) -> Self {
         Self { atom, chunk }
     }
-}
-
-pub enum ChunkReference<'a> {
-    //Not chopped
-    Center(&'a mut [Atom; CHUNK_LEN]),
-    //Chopped in two
-    Side([Option<[&'a mut Atom; HALF_CHUNK_LEN]>; 2]),
-    //Chopped in four
-    Corner([Option<[&'a mut Atom; QUARTER_CHUNK_LEN]>; 4]),
 }
 
 /// A deferred update message.
